requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-python-client==2.176.0
google-auth==2.40.3
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.4
python-docx==1.1.0
beautifulsoup4==4.12.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.6.1